refactor: add third-party and status apis + refactor routes.ts
